+++ Project-60-Typing_speed_game/README.md
@@ -1,10 +1,17 @@
-
 # Typing Speed Game
 
-This game tests the typing speed of user, it can 
-also be used to improve typing speed,or can be played for fun.
+This is a Python console-based typing speed game. It measures your typing speed (words per minute) and errors while typing a given prompt.
 
-## Screenshots
-![Getting Started](./type_speed_img.png)
-## Technologies used
-Python
+## Features
+- Accurate WPM calculation
+- Counts errors, including extra or missing words
+- Replayable without restarting
+- Clean and modular code for easy maintenance
+
+## How to Run
+1. Make sure you have Python installed (>=3.8 recommended).
+2. Open terminal/command prompt in the project directory.
+3. Run the script:
+   ```bash
+   python typing_speed_game.py
+4. Follow the on-screen instructions to start typing.
